requirements.txt
loguru
pyTelegramBotAPI
aiohttp
python-dotenv